gitlab: Create status with pr name
In gitlab the check names where showing up as a single one for every
pipelines.

This would cause issue when there is multiple pipelinerun running and
one of them fails. If other success it would override it and would show
as success in the Pipeline.

We now create each status like in github with the applicationName
/ OriginalPipelineName
